refactor(cli): thin CLI layer — move to beacon/cli/ and split by subcommand group
- Move beacon/core/cli/ → beacon/cli/; update shim import
- Rename underscore-prefixed cross-module names in display.py
  (_handle_soft_block → handle_soft_block, _print_doctor_summary → print_doctor_summary)
- Split cli/main.py into subcommand groups:
  - setup.py (setup)
  - sync.py (sync, delta, reset, update, clean, status)
  - contribute.py (contribute)
  - agent.py (agents, install, list)
  - adoption.py (adopt)
  - diagnostics.py (doctor)
- Keep cli/main.py as Click group + registration only
- Update all test imports and patches for new module paths
- Flip xfail markers in test_architecture.py for:
  - test_no_underscore_cross_module_imports
  - test_cli_handlers_have_no_io

Phase 8 of introduce-domain-layer. All 862 tests pass.
