+++ libs/beacon/src/beacon/cli/__init__.py
@@ -1,0 +1,1 @@
+"""Beacon CLI package — Click command handlers and output formatting."""
