--- libs/beacon/src/beacon/core/cli/__init__.py
@@ -1,1 +1,0 @@
-"""CLI package for Beacon."""
