+++ libs/beacon/tests/integration/test_contribute_agents.py
@@ -16,7 +16,7 @@ Test Cases:
 from pathlib import Path
 
 import pytest
-from beacon.cli import main
+from beacon.cli.main import main
 from click.testing import CliRunner
 
 # ---------------------------------------------------------------------------
