+++ libs/beacon/tests/integration/test_contribute_command.py
@@ -4,7 +4,7 @@ import subprocess
 from unittest.mock import MagicMock, patch
 
 import pytest
-from beacon.cli import main
+from beacon.cli.main import main
 from beacon.domains.artifact.skill import build_skills_paths
 from beacon.domains.contribution.contributor import (
     build_pr_body,
@@ -893,8 +893,8 @@ def test_contribute_auto_git_creates_pr(project_with_delta, tmp_path):
 
     runner = CliRunner()
     with (
-        patch("beacon.core.cli.main.check_warehouse_git_clean", return_value=None),
-        patch("beacon.core.cli.main.check_sync_state", return_value=None),
+        patch("beacon.cli.contribute.check_warehouse_git_clean", return_value=None),
+        patch("beacon.cli.contribute.check_sync_state", return_value=None),
         patch("beacon.domains.contribution.contributor.subprocess.run") as mock_run,
     ):
         mock_run.side_effect = [
@@ -934,8 +934,8 @@ def test_contribute_auto_git_fallback_when_push_fails(project_with_delta):
 
     runner = CliRunner()
     with (
-        patch("beacon.core.cli.main.check_warehouse_git_clean", return_value=None),
-        patch("beacon.core.cli.main.check_sync_state", return_value=None),
+        patch("beacon.cli.contribute.check_warehouse_git_clean", return_value=None),
+        patch("beacon.cli.contribute.check_sync_state", return_value=None),
         patch("beacon.domains.contribution.contributor.subprocess.run") as mock_run,
     ):
         mock_run.side_effect = [
@@ -958,8 +958,8 @@ def test_contribute_auto_git_fallback_when_gh_not_installed(project_with_delta):
 
     runner = CliRunner()
     with (
-        patch("beacon.core.cli.main.check_warehouse_git_clean", return_value=None),
-        patch("beacon.core.cli.main.check_sync_state", return_value=None),
+        patch("beacon.cli.contribute.check_warehouse_git_clean", return_value=None),
+        patch("beacon.cli.contribute.check_sync_state", return_value=None),
         patch("beacon.domains.contribution.contributor.subprocess.run") as mock_run,
     ):
         mock_run.side_effect = [
@@ -1204,8 +1204,8 @@ def test_default_auto_git_includes_untracked_in_pr(project_with_untracked):
 
     runner = CliRunner()
     with (
-        patch("beacon.core.cli.main.check_warehouse_git_clean", return_value=None),
-        patch("beacon.core.cli.main.check_sync_state", return_value=None),
+        patch("beacon.cli.contribute.check_warehouse_git_clean", return_value=None),
+        patch("beacon.cli.contribute.check_sync_state", return_value=None),
         patch("beacon.domains.contribution.contributor.subprocess.run") as mock_run,
     ):
         mock_run.side_effect = [
