+++ libs/beacon/tests/integration/test_contribute_noop.py
@@ -8,7 +8,7 @@ TDD Test Cases (9.1):
 """
 
 import pytest
-from beacon.cli import main
+from beacon.cli.main import main
 from click.testing import CliRunner
