+++ libs/beacon/tests/integration/test_e2e_happy_path.py
@@ -17,7 +17,7 @@ unit tests:
 
 import pytest
 import yaml
-from beacon.cli import main
+from beacon.cli.main import main
 from click.testing import CliRunner
 
 pytestmark = pytest.mark.integration
