+++ libs/beacon/tests/integration/test_install_flags.py
@@ -8,7 +8,7 @@ TDD Test Cases (10.4):
 """
 
 import pytest
-from beacon.cli import main
+from beacon.cli.main import main
 from click.testing import CliRunner
