+++ libs/beacon/tests/integration/test_reset_command.py
@@ -8,7 +8,7 @@ TDD Test Cases (8.3):
 """
 
 import pytest
-from beacon.cli import main
+from beacon.cli.main import main
 from click.testing import CliRunner
