+++ libs/beacon/tests/integration/test_status_command.py
@@ -5,7 +5,7 @@ Regression tests for:
 """
 
 import pytest
-from beacon.cli import main
+from beacon.cli.main import main
 from click.testing import CliRunner
