+++ libs/beacon/tests/integration/test_warehouse_git_check.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from unittest.mock import MagicMock, patch
 
 import pytest
-from beacon.cli import main
+from beacon.cli.main import main
 from beacon.domains.distribution.state import check_sync_state, write_sync_state
 from beacon.utils.git import check_warehouse_git_clean
 from click.testing import CliRunner
@@ -124,7 +124,7 @@ def test_sync_proceeds_when_warehouse_clean(connected_project):
 
     with (
         patch("subprocess.run") as mock_run,
-        patch("beacon.core.cli.main.check_warehouse_on_main_branch", return_value=None),
+        patch("beacon.cli.sync.check_warehouse_on_main_branch", return_value=None),
     ):
         mock_run.return_value = MagicMock(stdout="", returncode=0)
         result = runner.invoke(main, ["sync"])
@@ -154,7 +154,7 @@ def test_sync_dry_run_bypasses_git_check(connected_project):
     (warehouse / ".git").mkdir()
 
     # subprocess.run should NOT be called for the git check during dry-run
-    with patch("beacon.core.cli.main.check_warehouse_git_clean") as mock_check:
+    with patch("beacon.cli.sync.check_warehouse_git_clean") as mock_check:
         result = runner.invoke(main, ["sync", "--dry-run"])
 
     mock_check.assert_not_called()
@@ -206,7 +206,7 @@ def test_contribute_proceeds_when_warehouse_clean(connected_project_with_artifac
 
     with (
         patch("subprocess.run") as mock_run,
-        patch("beacon.core.cli.main.check_sync_state", return_value=None),
+        patch("beacon.cli.contribute.check_sync_state", return_value=None),
     ):
         mock_run.return_value = MagicMock(stdout="", returncode=0)
         result = runner.invoke(main, ["contribute", "knowledge/lesson.md"], input="y\n")
@@ -239,7 +239,7 @@ def test_contribute_dry_run_bypasses_git_check(connected_project_with_artifact):
     warehouse = connected_project_with_artifact["warehouse"]
     (warehouse / ".git").mkdir()
 
-    with patch("beacon.core.cli.main.check_warehouse_git_clean") as mock_check:
+    with patch("beacon.cli.contribute.check_warehouse_git_clean") as mock_check:
         result = runner.invoke(main, ["contribute", "knowledge/lesson.md", "--dry-run"])
 
     mock_check.assert_not_called()
@@ -635,7 +635,7 @@ def test_sync_proceeds_when_warehouse_on_main(connected_project):
 
     with (
         patch("subprocess.run") as mock_run,
-        patch("beacon.core.cli.main.check_warehouse_on_main_branch", return_value=None),
+        patch("beacon.cli.sync.check_warehouse_on_main_branch", return_value=None),
     ):
         mock_run.return_value = MagicMock(stdout="", returncode=0)
         result = runner.invoke(main, ["sync"])
@@ -662,7 +662,7 @@ def test_sync_dry_run_bypasses_branch_guard(connected_project):
     warehouse = connected_project["warehouse"]
     (warehouse / ".git").mkdir()
 
-    with patch("beacon.core.cli.main.check_warehouse_on_main_branch") as mock_branch:
+    with patch("beacon.cli.sync.check_warehouse_on_main_branch") as mock_branch:
         result = runner.invoke(main, ["sync", "--dry-run"])
 
     mock_branch.assert_not_called()
