+++ libs/beacon/tests/integration/test_warehouse_init.py
@@ -4,7 +4,7 @@ import subprocess
 from pathlib import Path
 
 import pytest
-from beacon.cli import main
+from beacon.cli.main import main
 from click.testing import CliRunner
